+++ src/main/java/com/gdmc/httpinterfacemod/handlers/CommandHandler.java
@@ -1,13 +1,17 @@
 package com.gdmc.httpinterfacemod.handlers;
 
+import com.google.gson.JsonArray;
+import com.google.gson.JsonObject;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.sun.net.httpserver.Headers;
 import com.sun.net.httpserver.HttpExchange;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.server.MinecraftServer;
 
-import java.io.*;
-import java.util.ArrayList;
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
@@ -31,32 +35,33 @@ public class CommandHandler extends HandlerBase {
 
         CommandSourceStack cmdSrc = createCommandSource("GDMC-CommandHandler", dimension);
 
-        List<String> outputs = new ArrayList<>();
+        JsonArray returnValues = new JsonArray();
         for (String command: commands) {
             if (command.length() == 0) {
                 continue;
             }
             // requests to run the actual command execution on the main thread
-            CompletableFuture<String> cfs = CompletableFuture.supplyAsync(() -> {
+            CompletableFuture<JsonObject> cfs = CompletableFuture.supplyAsync(() -> {
                 try {
-                    return "" + mcServer.getCommands().getDispatcher().execute(command, cmdSrc);
+                    int commandStatus = mcServer.getCommands().getDispatcher().execute(command, cmdSrc);
+                    return instructionStatus(
+                        commandStatus != 0,
+                        commandStatus != 1 && commandStatus != 0 ? String.valueOf(commandStatus) : null
+                    );
                 } catch (CommandSyntaxException e) {
-                    return e.getMessage();
+                    return instructionStatus(false, e.getMessage());
                 }
             }, mcServer);
 
             // block this thread until the above code has run on the main thread
-            String result = cfs.join();
-            outputs.add(result);
+            returnValues.add(cfs.join());
         }
 
         // Response headers
         Headers responseHeaders = httpExchange.getResponseHeaders();
         setDefaultResponseHeaders(responseHeaders);
-        setResponseHeadersContentTypePlain(responseHeaders);
 
         // body
-        String responseString = String.join("\n", outputs);
-        resolveRequest(httpExchange, responseString);
+        resolveRequest(httpExchange, returnValues.toString());
     }
 }