+++ src/main/java/com/gdmc/httpinterfacemod/handlers/EntitiesHandler.java
@@ -1,6 +1,5 @@
 package com.gdmc.httpinterfacemod.handlers;
 
-import com.gdmc.httpinterfacemod.utils.JsonTagVisitor;
 import com.gdmc.httpinterfacemod.utils.TagMerger;
 import com.google.gson.*;
 import com.mojang.brigadier.StringReader;
@@ -23,11 +22,13 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.Vec3;
 
-import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.util.*;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
 
 public class EntitiesHandler extends HandlerBase {
 
@@ -71,30 +72,22 @@ public class EntitiesHandler extends HandlerBase {
 			throw new HttpException(message, 400);
 		}
 
-		// Check if clients wants a response in a JSON format. If not, return response in plain text.
-		Headers requestHeaders = httpExchange.getRequestHeaders();
-		String acceptHeader = getHeader(requestHeaders, "Accept", "*/*");
-		boolean returnJson = hasJsonTypeInHeader(acceptHeader);
-
 		String method = httpExchange.getRequestMethod().toLowerCase();
 
-		String contentTypeHeader = getHeader(requestHeaders, "Content-Type", "*/*");
-		boolean parseRequestAsJson = hasJsonTypeInHeader(contentTypeHeader);
-
-		String responseString;
+		JsonArray response;
 
 		switch (method) {
 			case "put" -> {
-				responseString = putEntitiesHandler(httpExchange.getRequestBody(), parseRequestAsJson, returnJson);
+				response = putEntitiesHandler(httpExchange.getRequestBody());
 			}
 			case "get" -> {
-				responseString = getEntitiesHandler(returnJson);
+				response = getEntitiesHandler();
 			}
 			case "delete" -> {
-				responseString = deleteEntitiesHandler(httpExchange.getRequestBody(), parseRequestAsJson, returnJson);
+				response = deleteEntitiesHandler(httpExchange.getRequestBody());
 			}
 			case "patch" -> {
-				responseString = patchEntitiesHandler(httpExchange.getRequestBody(), parseRequestAsJson, returnJson);
+				response = patchEntitiesHandler(httpExchange.getRequestBody());
 			}
 			default -> {
 				throw new HttpException("Method not allowed. Only PUT, GET, DELETE and PATCH requests are supported.", 405);
@@ -104,72 +97,46 @@ public class EntitiesHandler extends HandlerBase {
 		// Response headers
 		Headers responseHeaders = httpExchange.getResponseHeaders();
 		setDefaultResponseHeaders(responseHeaders);
-		if (returnJson) {
-			setResponseHeadersContentTypeJson(responseHeaders);
-		} else {
-			setResponseHeadersContentTypePlain(responseHeaders);
-		}
 
-		resolveRequest(httpExchange, responseString);
+		resolveRequest(httpExchange, response.toString());
 	}
 
 	/**
 	 * @param requestBody request body of entity summon instructions
-	 * @param parseRequestAsJson if true, treat input as JSON
-	 * @param returnJson if true, return result in JSON format
 	 * @return summon results
 	 */
-	private String putEntitiesHandler(InputStream requestBody, boolean parseRequestAsJson, boolean returnJson) {
+	private JsonArray putEntitiesHandler(InputStream requestBody) {
 		CommandSourceStack cmdSrc = createCommandSource("GDMC-EntitiesHandler", dimension).withPosition(new Vec3(x, y, z));
 		ServerLevel serverLevel = getServerLevel(dimension);
 
-		ArrayList<String> returnValues = new ArrayList<>();
-		if (parseRequestAsJson) {
-			JsonArray entityDescriptionList;
+		JsonArray returnValues = new JsonArray();
+		JsonArray entityDescriptionList;
+		try {
+			entityDescriptionList = JsonParser.parseReader(new InputStreamReader(requestBody)).getAsJsonArray();
+		} catch (JsonSyntaxException jsonSyntaxException) {
+			throw new HttpException("Malformed JSON: " + jsonSyntaxException.getMessage(), 400);
+		}
+
+		for (JsonElement entityDescription : entityDescriptionList) {
+			JsonObject json = entityDescription.getAsJsonObject();
+
+			SummonEntityInstruction summonEntityInstruction;
 			try {
-				entityDescriptionList = JsonParser.parseReader(new InputStreamReader(requestBody)).getAsJsonArray();
-			} catch (JsonSyntaxException jsonSyntaxException) {
-				throw new HttpException("Malformed JSON: " + jsonSyntaxException.getMessage(), 400);
+				summonEntityInstruction = new SummonEntityInstruction(json, cmdSrc);
+			} catch (CommandSyntaxException e) {
+				returnValues.add(instructionStatus(false, e.getMessage()));
+				continue;
 			}
-
-			for (JsonElement entityDescription : entityDescriptionList) {
-				JsonObject json = entityDescription.getAsJsonObject();
-
-				SummonEntityInstruction summonEntityInstruction;
-				try {
-					summonEntityInstruction = new SummonEntityInstruction(json, cmdSrc);
-				} catch (CommandSyntaxException e) {
-					returnValues.add(e.getMessage());
-					continue;
-				}
-				returnValues.add(summonEntityInstruction.summon(serverLevel));
-			}
-		} else {
-			List<String> inputList = new BufferedReader(new InputStreamReader(requestBody)).lines().toList();
-			for (String inputSummonInstruction : inputList) {
-				SummonEntityInstruction summonEntityInstruction;
-				try {
-					summonEntityInstruction = new SummonEntityInstruction(inputSummonInstruction, cmdSrc);
-				} catch (CommandSyntaxException e) {
-					returnValues.add(e.getMessage());
-					continue;
-				}
-				returnValues.add(summonEntityInstruction.summon(serverLevel));
-			}
+			returnValues.add(summonEntityInstruction.summon(serverLevel));
 		}
 
-		// Set response as a list of "1" (entity was placed), "0" (entity was not placed) or an exception string if something went wrong.
-		if (returnJson) {
-			return new Gson().toJson(returnValues);
-		}
-		return String.join("\n", returnValues);
+		return returnValues;
 	}
 
 	/**
-	 * @param returnJson if true, return resposne in JSON formatted string
 	 * @return list of entity information
 	 */
-	private String getEntitiesHandler(boolean returnJson) {
+	private JsonArray getEntitiesHandler() {
 
 		// Calculate boundaries of area of blocks to gather information on.
 		int xOffset = x + dx;
@@ -188,62 +155,44 @@ public class EntitiesHandler extends HandlerBase {
 
 		List<Entity> entityList = level.getEntities(null, new AABB(xMin, yMin, zMin, xMax, yMax, zMax));
 
-		if (returnJson) {
-			// Create a JsonArray with JsonObject, each contain a key-value pair for
-			// the x, y, z position, the block ID, the block state (if requested and available)
-			// and the block entity data (if requested and available).
-			JsonArray jsonArray = new JsonArray();
-			for (Entity entity : entityList) {
-				String entityId = entity.getEncodeId();
-				if (entityId == null) {
-					continue;
-				}
-				JsonObject json = new JsonObject();
-				json.addProperty("uuid", entity.getStringUUID());
-				if (includeData) {
-					json.add("data", getEntityDataAsJsonObject(entity));
-				}
-				jsonArray.add(json);
-			}
-			return new Gson().toJson(jsonArray);
-		}
-
-		ArrayList<String> responseList = new ArrayList<>();
+		// Create a JsonArray with JsonObject, each contain a key-value pair for
+		// the x, y, z position, the block ID, the block state (if requested and available)
+		// and the block entity data (if requested and available).
+		JsonArray returnList = new JsonArray();
 		for (Entity entity : entityList) {
 			String entityId = entity.getEncodeId();
 			if (entityId == null) {
 				continue;
 			}
-			responseList.add(entity.getStringUUID() + " " + getEntityDataAsStr(entity));
+			JsonObject json = new JsonObject();
+			json.addProperty("uuid", entity.getStringUUID());
+			if (includeData) {
+				json.addProperty("data", getEntityDataAsStr(entity));
+			}
+			returnList.add(json);
 		}
-		return String.join("\n", responseList);
+		return returnList;
 	}
 
 	/**
 	 * @param requestBody request body of entity removal instructions
-	 * @param parseRequestAsJson if true, treat input as JSON
-	 * @param returnJson if true, return result in JSON format
 	 * @return entity removal results
 	 */
-	private String deleteEntitiesHandler(InputStream requestBody, boolean parseRequestAsJson, boolean returnJson) {
+	private JsonArray deleteEntitiesHandler(InputStream requestBody) {
 
 		ServerLevel level = getServerLevel(dimension);
 
 		List<String> entityUUIDToBeRemoved;
 
-		List<String> returnValues = new ArrayList<>();
+		JsonArray returnValues = new JsonArray();
 
-		if (parseRequestAsJson) {
-			JsonArray jsonListUUID;
-			try {
-				jsonListUUID = JsonParser.parseReader(new InputStreamReader(requestBody)).getAsJsonArray();
-			} catch (JsonSyntaxException jsonSyntaxException) {
-				throw new HttpException("Malformed JSON: " + jsonSyntaxException.getMessage(), 400);
-			}
-			entityUUIDToBeRemoved = Arrays.asList(new Gson().fromJson(jsonListUUID, String[].class));
-		} else {
-			entityUUIDToBeRemoved = new BufferedReader(new InputStreamReader(requestBody)).lines().toList();
+		JsonArray jsonListUUID;
+		try {
+			jsonListUUID = JsonParser.parseReader(new InputStreamReader(requestBody)).getAsJsonArray();
+		} catch (JsonSyntaxException jsonSyntaxException) {
+			throw new HttpException("Malformed JSON: " + jsonSyntaxException.getMessage(), 400);
 		}
+		entityUUIDToBeRemoved = Arrays.asList(new Gson().fromJson(jsonListUUID, String[].class));
 
 		for (String stringUUID : entityUUIDToBeRemoved) {
 			if (stringUUID.length() == 0) {
@@ -253,107 +202,62 @@ public class EntitiesHandler extends HandlerBase {
 			try {
 				entityToBeRemoved = level.getEntity(UUID.fromString(stringUUID));
 			} catch (IllegalArgumentException e) {
-				returnValues.add("0");
+				returnValues.add(instructionStatus(false, e.getMessage()));
 				continue;
 			}
 
 			if (entityToBeRemoved != null) {
 				if (entityToBeRemoved.isRemoved()) {
-					returnValues.add("0");
+					returnValues.add(instructionStatus(false));
 				} else {
 					entityToBeRemoved.remove(Entity.RemovalReason.DISCARDED);
-					returnValues.add("1");
+					returnValues.add(instructionStatus(true));
 				}
 				continue;
 			}
-			returnValues.add("0");
+			returnValues.add(instructionStatus(false));
 		}
-
-		if (returnJson) {
-			return new Gson().toJson(returnValues);
-		}
-		return String.join("\n", returnValues);
+		return returnValues;
 	}
 
 	/**
 	 * @param requestBody request body of entity patch instructions
-	 * @param parseRequestAsJson if true, treat input as JSON
-	 * @param returnJson if true, return result in JSON-formatted string
 	 * @return entity patch status results
 	 */
-	private String patchEntitiesHandler(InputStream requestBody, boolean parseRequestAsJson, boolean returnJson) {
+	private JsonArray patchEntitiesHandler(InputStream requestBody) {
 
 		ServerLevel level = getServerLevel(dimension);
 
-		List<String> returnValues = new ArrayList<>();
+		JsonArray returnValues = new JsonArray();
 
-		if (parseRequestAsJson) {
-			JsonArray jsonList;
+		JsonArray jsonList;
+		try {
+			jsonList = JsonParser.parseReader(new InputStreamReader(requestBody)).getAsJsonArray();
+		} catch (JsonSyntaxException jsonSyntaxException) {
+			throw new HttpException("Malformed JSON: " + jsonSyntaxException.getMessage(), 400);
+		}
+		for (JsonElement entityDescription : jsonList) {
+			JsonObject json = entityDescription.getAsJsonObject();
+			PatchEntityInstruction patchEntityInstruction;
 			try {
-				jsonList = JsonParser.parseReader(new InputStreamReader(requestBody)).getAsJsonArray();
-			} catch (JsonSyntaxException jsonSyntaxException) {
-				throw new HttpException("Malformed JSON: " + jsonSyntaxException.getMessage(), 400);
+				patchEntityInstruction = new PatchEntityInstruction(json);
+			} catch (IllegalArgumentException | CommandSyntaxException | UnsupportedOperationException e) {
+				returnValues.add(instructionStatus(false, e.getMessage()));
+				continue;
 			}
-			for (JsonElement entityDescription : jsonList) {
-				JsonObject json = entityDescription.getAsJsonObject();
-				PatchEntityInstruction patchEntityInstruction;
-				try {
-					patchEntityInstruction = new PatchEntityInstruction(json);
-				} catch (IllegalArgumentException | CommandSyntaxException e) {
-					returnValues.add(e.getMessage());
+			try {
+				if (patchEntityInstruction.applyPatch(level)) {
+					returnValues.add(instructionStatus(true));
 					continue;
 				}
-				try {
-					if (patchEntityInstruction.applyPatch(level)) {
-						returnValues.add("1");
-						continue;
-					}
-				} catch (ReportedException e) {
-					returnValues.add(e.getMessage());
-					continue;
-				}
-				returnValues.add("0");
+			} catch (ReportedException e) {
+				returnValues.add(instructionStatus(false, e.getMessage()));
+				continue;
 			}
-		} else {
-			List<String> textList = new BufferedReader(new InputStreamReader(requestBody)).lines().toList();
-			for (String entityDescription : textList) {
-				PatchEntityInstruction patchEntityInstruction;
-				try {
-					patchEntityInstruction = new PatchEntityInstruction(entityDescription);
-				} catch (IllegalArgumentException | CommandSyntaxException e) {
-					returnValues.add(e.getMessage());
-					continue;
-				}
-				try {
-					if (patchEntityInstruction.applyPatch(level)) {
-						returnValues.add("1");
-						continue;
-					}
-				} catch (ReportedException e) {
-					returnValues.add(e.getMessage());
-					continue;
-				}
-				returnValues.add("0");
-			}
+			returnValues.add(instructionStatus(false));
 		}
 
-		if (returnJson) {
-			return new Gson().toJson(returnValues);
-		}
-		return String.join("\n", returnValues);
-	}
-
-	private JsonObject getEntityDataAsJsonObject(Entity entity) {
-		JsonObject json = new JsonObject();
-		CompoundTag tags = entity.serializeNBT();
-		if (tags != null) {
-			String tagAsJsonString = (new JsonTagVisitor()).visit(tags);
-			JsonObject tagsAsJsonObject = JsonParser.parseString(tagAsJsonString).getAsJsonObject();
-			if (tagsAsJsonObject != null) {
-				return tagsAsJsonObject;
-			}
-		}
-		return json;
+		return returnValues;
 	}
 
 	private String getEntityDataAsStr(Entity entity) {
@@ -373,30 +277,15 @@ public class EntitiesHandler extends HandlerBase {
 		private Vec3 entityPosition;
 		private CompoundTag entityData;
 
-		SummonEntityInstruction(JsonObject inputData, CommandSourceStack commandSourceStack) throws CommandSyntaxException {
+		SummonEntityInstruction(JsonObject summonInstructionInput, CommandSourceStack commandSourceStack) throws CommandSyntaxException {
 			String positionArgumentString = "";
-			if (inputData.has("x") && inputData.has("y") && inputData.has("z")) {
-				positionArgumentString = inputData.get("x").getAsString() + " " + inputData.get("y").getAsString() + " " + inputData.get("z").getAsString();
-				inputData.remove("x");
-				inputData.remove("y");
-				inputData.remove("z");
-			} else if (inputData.has("Pos") && inputData.get("Pos").isJsonArray()) {
-				JsonArray positionInputData = inputData.get("Pos").getAsJsonArray();
-				if (positionInputData.size() == 3) {
-					positionArgumentString = "%s %s %s".formatted(
-						positionInputData.get(0).getAsString(),
-						positionInputData.get(1).getAsString(),
-						positionInputData.get(2).getAsString()
-					);
-				}
+			if (summonInstructionInput.has("x") && summonInstructionInput.has("y") && summonInstructionInput.has("z")) {
+				positionArgumentString = summonInstructionInput.get("x").getAsString() + " " + summonInstructionInput.get("y").getAsString() + " " + summonInstructionInput.get("z").getAsString();
 			}
-			String entitySummonArgumentString = inputData.has("id") ? inputData.get("id").getAsString() : "";
+			String entityIDString = summonInstructionInput.has("id") ? summonInstructionInput.get("id").getAsString() : "";
+			String entityDataString = summonInstructionInput.has("data") ? summonInstructionInput.get("data").getAsString() : "";
 
-			parse(positionArgumentString + " " + entitySummonArgumentString + " " + inputData, commandSourceStack);
-		}
-
-		SummonEntityInstruction(String inputData, CommandSourceStack commandSourceStack) throws CommandSyntaxException {
-			parse(inputData, commandSourceStack);
+			parse(positionArgumentString + " " + entityIDString + " " + entityDataString, commandSourceStack);
 		}
 
 		private void parse(String inputData, CommandSourceStack commandSourceStack) throws CommandSyntaxException {
@@ -409,15 +298,19 @@ public class EntitiesHandler extends HandlerBase {
 			sr.skip();
 
 			try {
-				entityData = TagParser.parseTag(sr.getRemaining());
+				String entityDataString = sr.getRemaining();
+				if (entityDataString.isBlank()) {
+					entityDataString = "{}";
+				}
+				entityData = TagParser.parseTag(entityDataString);
 			} catch (StringIndexOutOfBoundsException e) {
 				entityData = new CompoundTag();
 			}
 		}
 
-		public String summon(ServerLevel level) {
+		public JsonObject summon(ServerLevel level) {
 			if (!Level.isInSpawnableBounds(new BlockPos(entityPosition))) {
-				return "Position is not in spawnable bounds";
+				return instructionStatus(false, "Position is not in spawnable bounds");
 			}
 
 			entityData.putString("id", entityResourceLocation.toString());
@@ -427,16 +320,16 @@ public class EntitiesHandler extends HandlerBase {
 				return _entity;
 			});
 			if (entity == null) {
-				return "Entity could not be spawned";
+				return instructionStatus(false, "Entity could not be spawned");
 			}
 			entity.checkDespawn();
 			if (entity.isRemoved()) {
-				return "Entity was removed right after spawn for reason: " + entity.getRemovalReason();
+				return instructionStatus(false, "Entity was removed right after spawn for reason: " + entity.getRemovalReason());
 			}
 			if (!level.tryAddFreshEntityWithPassengers(entity)) {
-				return "Entity with this UUID already exists";
+				return instructionStatus(false, "Entity with this UUID already exists");
 			}
-			return entity.getStringUUID();
+			return instructionStatus(true, entity.getStringUUID());
 		}
 
 	}
@@ -448,15 +341,9 @@ public class EntitiesHandler extends HandlerBase {
 		private final UUID uuid;
 		private final CompoundTag patchData;
 
-		PatchEntityInstruction(JsonObject inputData) throws IllegalArgumentException, CommandSyntaxException {
+		PatchEntityInstruction(JsonObject inputData) throws IllegalArgumentException, CommandSyntaxException, UnsupportedOperationException {
 			uuid = UUID.fromString(inputData.get("uuid").getAsString());
-			patchData = TagParser.parseTag(inputData.get("data").toString());
-		}
-
-		PatchEntityInstruction(String inputData) throws IllegalArgumentException, CommandSyntaxException {
-			StringReader sr = new StringReader(inputData);
-			uuid = UUID.fromString(sr.readStringUntil(' '));
-			patchData = TagParser.parseTag(sr.getRemaining());
+			patchData = TagParser.parseTag(inputData.get("data").getAsString());
 		}
 
 		public boolean applyPatch(ServerLevel level) throws ReportedException {